implemented store in watchlist

+++ kino-app/src/app/movies/watchlist/watchlist.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Subscription } from 'rxjs';
+import { Subscription, Observable } from 'rxjs';
+import { User } from 'src/app/types';
 import { WatchlistService } from './watchlist.service';
 
 
@@ -10,20 +11,31 @@ import { WatchlistService } from './watchlist.service';
   styleUrls: ['./watchlist.component.scss']
 })
 export class WatchlistComponent implements OnInit, OnDestroy {
-  userId = 1;
+  user$: Observable<User | undefined>;
+  userId: number | undefined = undefined;
+
   private subscription = new Subscription()
 
-  // zmienić mockowany userId na store
+  // zmienić mockowany userId na store, nie dawać opcji wejścia jak niezalogowany
 
   constructor(
     private service : WatchlistService,
-    // private store : Store jeszcze nie wiem na co on tu jest ale będzie sobie brał userId
-  ) {}
+    private store : Store<{ userData: { user?: User } }>
+  ) {
+    this.user$ = this.store.select(state => state.userData.user);
+    this.user$.subscribe(userData => {
+      this.userId = userData?.id
+    });
+  }
 
   userWatchlist$ = this.service.userWatchlist$
 
-  remove(id : number, movie : number) { 
-    this.service.removeFromWatchlist(id, movie)
+
+
+  remove(id : number | undefined, movie : number) {
+    if (id) {
+      this.service.removeFromWatchlist(id, movie)
+    } 
   }
   ngOnInit(): void {
     this.service.getWatchlistMovies(this.userId)
